Update event length filtering to accept None as default for max window
Previously max window was being set the number of elements in the initial condition which is implicitly moving the max window to a large enough value that it wouldn't matter. Now max window is explicit in that it will not try to filter for max window if it's the default value of None. Conditions for if max and min window are present are now explicit as well so as to communicate the true conditions to enable event length filtering.