Import C based debounce, and fall back to Python method if there's an ImportError
I couldn't get the package to install without doing this. I kept getting an error that it couldn't find debounce.c because it was looking for the .pyd file which isn't part of the repo. Not sure if this is the best way to address this, but I'm out of ideas.