Change all keyword defaults to None.
Property defaults are now set in the property definition. This is to allow the skip_check decorator to only have to determine if attributes are None, or not None to work.

Skip check needs to check the actual attribute and not the property. This is confusing and not ideal, but it's ok for now. Maybe default checks need to be done in the method? That way the logic for each will be specific, however there will be alot of code reuse that way.