Refactor API
Remove transformation methods from the Events class and move them to their own module. 

Transformations are now set up in a funcational approach to make it easier to extend by users and to allow more control as to the order they are applied.

Add find_events() that returns an Event object. The Event object has been refactored to fit the new API.